network death lock check (#36)
* fix window build from MIN_WATERMARK

* Client tcp set R/W timeout

* async thread named

* replace `tokio::time::sleep` with `async_sleep`

* network client codereview

* panic format

* thread info collect

* worker add web ui server

* network client log switch

* client add batch_size check

* client send element block log

* set network client thread pool size

* update version 0.2.5

* udpate connector version 0.2.5

* add num of cpu cores log

* network server log enable api

* network server batch send size log

* network client debug log

* netowrk batch request add batch_id field

* network response serde

* rename ResonseCode fn `err` to `end`

* network server batch response

* network client request timeout & end status mock

* remove unused code

* network package serialize length check

* network request&response serde check

* channel bound_base log

* network disable log as default

* element `Record` serde check

* add startup_time & start_number metric

* fix startup_number named

* starup arguments gauge

+++ rlink/src/runtime/context.rs
@@ -63,8 +63,8 @@ pub(crate) struct Context {
     /// on yarn arg
     pub yarn_manager_main_class: String,
     pub worker_process_path: String,
-    pub memory_mb: usize,
-    pub v_cores: usize,
+    pub memory_mb: u32,
+    pub v_cores: u32,
 }
 
 impl Context {
@@ -82,8 +82,8 @@ impl Context {
         dashboard_path: String,
         yarn_manager_main_class: String,
         worker_process_path: String,
-        memory_mb: usize,
-        v_cores: usize,
+        memory_mb: u32,
+        v_cores: u32,
     ) -> Self {
         Context {
             application_name,
@@ -169,12 +169,12 @@ impl Context {
                     let worker_process_path = parse_arg("worker_process_path")?;
 
                     let memory_mb = parse_arg("memory_mb")?;
-                    let memory_mb = usize::from_str(memory_mb.as_str()).map_err(|_e| {
+                    let memory_mb = u32::from_str(memory_mb.as_str()).map_err(|_e| {
                         anyhow!("parse `memory_mb`=`{}` to usize error", memory_mb)
                     })?;
 
                     let v_cores = parse_arg("v_cores")?;
-                    let v_cores = usize::from_str(v_cores.as_str())
+                    let v_cores = u32::from_str(v_cores.as_str())
                         .map_err(|_e| anyhow!("parse `v_cores`=`{}` to usize error", v_cores))?;
 
                     (
